climb: Rename IsClimbing to IsPistonExtended.
Add IsPistonExtended and IsWinchOut flags for those conditions.
IsPistonExtended replaces old IsClimbing flag.

IsClimbing assumed that climbing only involved the piston.  If the winch
has released cable at all, then the robot is climbing.

If the operator extends the piston, winches cable out, but isn't hanging
on the bar, then retracting the piston will drop slack cable onto the
robot.  The robot shall remain in a climbing state even if the piston is
retracted in this situation.

If climbing ever starts, set climb flag and leave it set.  We have no
means to measure if operator retracted the climb mechanism properly, and
operating, say, the conveyor with slack cable may cause damage.

Let `IsClimbing = IsClimbing || IsWinchOut`, since it's slack cable that
is of more concern than the piston.

The operator can accidently extend the piston and retract it if no cable
is let out, it's safe to assume the robot isn't in a climbing state.

Signed-off-by: Justin Charette <charetjc@gmail.com>