Rationale commit
Currently we use Shuffleboard for our dashboard on the driver station. The Driving network table contains all the data that we can push (which does not include from the SendableChooser and the Limelight camera feed) that the drive team needs to run a match. This table should be reserved for this purpose alone, as currently we simply push any data we may need during testing with SmartDashboard. As I understand, Shuffleboard allows us to use data from other tables on the same page as the ones specifically for it, and we can set up a nice view that will remain persistent between reboots of the robot and not be cluttered with data the drive team does not need.